Safe#delete will not destroy personal files
I made an in-place safe that contained my entire home directory, and
realized that if I unthinkingly called `#delete` on the safe I would be
doing `rm -rf ~`. Not cool. Now `Safe#delete` will first destroy all
the sia-specific files in the safe (encrypted files, index file, and
salt file), then the safe directory will only be deleted if it is empty.
So if there's anything unexpected in there it won't get deleted.