Introducing portable safes
Portable safes are similar to how in-place safes used to behave. Now,
in-place safes will encrypt files in-place as before (only adding a file
extension), but they are no longer required to be within the safe_dir.
Portable safes enforce that the clearfile must be within the safe_dir,
but do not encrypt files in-place. Both options can be combined.